requirements.txt
fastapi
uvicorn
python-dotenv
langchain-community
openweathermap
requests
langchain

langchain-openai
langchain-classic
langchain-core